Jeff Carneal <jeff@apex.net>
Modified lowercase/nospace functions to work either
before or after processing the request.  If after,
the server will retry the request with the new values.

Added function rad_getpass() to auth.c.

This will all be moved to a module after we have the
ability to auth -> fail -> lower -> auth again in
module processing.