Add the default_user_profile configuration directive and support for the User-Profile attribute in the sql module. The default_profile is found in SQL by group membership. That means that this profile must be a member of at least one group which will contain the corresponding check and reply items. This profile will be queried in the authorize section for every user. The point is to assign all users a default profile without having to manually add each one to a group that will contain the profile. The SQL module will also honor the User-Profile attribute. This attribute can be set anywhere in the authorize section (ie the users file). It is found exactly as the default profile is found. If it is set then it will *overwrite* the default profile setting. The idea is to select profiles based on checks on the incoming packets, not on user group membership. For example: -- users file -- DEFAULT       Service-Type == Outbound-User, User-Profile := "outbound" DEFAULT       Service-Type == Framed-User, User-Profile := "framed"
By default the default_user_profile is not set. We also add the
query_on_not_found configuration directive which determines if we will
query the default_user_profile or the User-Profile if the user is not found.
If the profile is found then we consider the user found. That way we can still
maintain the old module behaviour to look for a DEFAULT entry if the user was
not found. But now that is configurable and we don't need to do extra queries
if the user is not found. By default this is set to 'no'.


Also change references to RFC 2138 to RFC 2865 in the dictionary man page.