Add a one line patch from Martin Seine (http://lists.cistron.nl/pipermail/freeradius-devel/2004-March/006897.html)
Keep the number of retransmissions for each outstanding request.
If we do a retransmission we set the retransmit time by the formula:
       if (r->retrans_num > 20)
               r->retrans = now + 70;
       else
               r->retrans = now + 3 + (3 * r->retrans_num);

The above can help if the radius server is just overloaded and can't respond quickly enough
to the requests. Even if it is down, sending packets every 70secs instead of 3secs is
generally a good idea.