Don't double-escape strings in pairread()
If 'value' is a double-quoted string, then gettoken() already
escaped \n -> 0x0a.  Calling pairmake() with value does the
escaping again, which is wrong.  Try this with:

DEFAULT
	Filter-Id := "foo\nbar\\n"

gettoken() converts the first \n to 0x0a, and the \\ to \.
pairmake() then converts the last \n to 0x0a, leaving 2 0x0a's
in the string, rather than on 0x0a, and another \n.

I've also added handlers in pairread() for single quoted strings,
which didn't previously exist.