PEAP & TTLS support for session resumption.
It works (sort of).  it doesn't cache the original inner username,
so the user name in the reply is wrong (i.e. anonymous).  It SHOULD
cache a lot more things, like CUI.  This list could also be
configurable.

It also needs to run the resumption stuff through a virtual server
again, to see if the user is still authorized.

OR, have an attribute that's added to the request to mark it as
session resumption, and then any post-auth policy can key off of that,
and do more stuff