Move socket listening code to plugin modules.
This makes the server core a fair bit simpler, and allows for
the creation of many new kinds of listening sockets.

The next step is to delete the RAD_LISTEN_TYPE_* names, as they're
no longer really needed.

The goal is to move much of src/main/event.c into the plugin modules,
so that the server core isn't are of the underlying protocols.