start sending Status-Server messages earlier
We now start sending them at the START of the zombie period,
which means that a "live" home server is marked "dead" less often.

i.e. an upstream proxy is responsible for realms A && B.  We're
currently proxying packets for realm A only... and their upstream
is down.  If our upstream doesn't respond, we might think that
it's down, when really it's being stupid.  So, we start pinging
it with Status-Server.  If it responds, we keep sending it packets,
including packets for realm B.