Change where we do suid up/down.
If the server starts as root, but it supposed to run as another
user, we want to *temporarily* drop permissions very early.  Then,
when binding to privileged sockets, we re-gain permissions.
Once all of the sockets are open, we drop them permanently.

However, if we suid up for *all* sockets, then the control socket
will be created as root, rather than as the unprivileged user.
To fix that, we put suid up/down just around the 2 calls that
need it.