Fix broken EAP-TLS (bug introduced 2008/08/24 by b51a3a82)
Thu, 29 Jan 2009 16:31:10 +0100
To: aland <aland@deployingradius.com>
CC: Axel Tillequin <axel.tillequin@eads.net>, FreeRadius developers mailing list <freeradius-devel@lists.freeradius.org>

Hi,

As explained in previous mails of the thread, FreeRadius EAP-TLS support
is broken (the EAP encapsulated TLS ChangeCipherSpec and TLS Finished
messages are not sent). Bisecting the issue led me here:

    commit b51a3a82edb797f5d0a2758bd1a38359d6f66803
    Author: Alan T. DeKok <aland@freeradius.org>
    Date:   Sun Aug 24 10:04:55 2008 +0200

        Clean up debug && log messages

AFAICT, the test that prevented eaptls_ack_handler() to return
EAPTLS_SUCCESS *before* flushing remaining local messages
(i.e. returning EAPTLS_REQUEST so that they be sent to the peer to
complete the TLS handshake) was removed in that commit.

The patch below is against current git tree. With Axel, we tested the
fix with 2.1.3: it corrects the issue.

Cheers,

a+

Signed-off-by: Arnaud Ebalard <arno@natisbad.org>
Tested-by: Axel Tillequin <axel.tillequin@gmail.com>