Cleaned up code based on report of crash.
Added additional notes on coa

moved "check for reply" in received_proxy_response to earlier,
as it shouldn't be done later.

Simplified check for CoA divorce.  This is the code that crashed
(still no idea why), but the new code should do the intended checks,
without the un-intended side effects