Allow fast clients to send packets, too
This is for performance testing.  For speed, the server doesn't have
locks all over the place.  However, this means that when it runs threaded,
the client can receive the response, and send a new one using the same
ID, before the child thread has a chance to update child_state.

This work-around lets everyone be happy