Allow operations OTHER than == to work for Packet-Src-IP-Address
... and associated virtual attributes.

The issue is that the paircompare_register'd functions return 0
for match, and 1 for didn't match.  This is wrong.  They should just
return the results of the comparison.  And then radius_callback_compare
should check the results of the comparison against the operators,
to see if the CONDITION succeeded or failed.