"perl -MExtUtils::Embed -e ldopts" LIES to us
So... check if the lying liar is lying.  If so, complain, and
refuse to build the module.

i.e. it says "Use -lperl to link!"  But there's no "libperl.so"
on the system.  And the upstream developers don't see a problem
with Perl lying to the applications.  This means that we have to
check for liars, and to work around bugs in other peoples software