Copy the User-Name by value
The previous method treated the User-Name as a string, and parsed it
to create the User-Name for the reply.  However... if that happens,
it SHOULD print the User-Name to a string, and then parse that.
That way things like 'FOO\tbar' will get escaped to 'FOO\\tbar',
and therefore parsed properly.

Or, we could just copy the contents verbatim, which is what we did