Preliminary implementation of DHCP relay
For now, it can only relay packets if giaddr = 0.
If giaddr != 0, we are required to cache the packet by XID/MAC, so that
we can capture the response on the DHCP port, look up the XID/MAC, and
respond to the client.

We will also need to update the fr_dhcp_encode function, which can
currently only respond to clients, and not originate packets.