Jumbo patch to clean up socket handling
The transport protocol code is now more merged, and the "fr_tcp_radius_t"
structure and API are deleted.  The resulting code is simpler and smaller.

Cleaned up how the listeners and even handlers deal with sockets.
The proxy sockets are now pushed into the proxy packet list, and are
not used in the main listener list.

The proxy packet list now deals with src/dst, and not just destination.