Ensure that there is a cleanup event for proxied packets
If there was no reply, clean up, reject, etc. the request.

This doesn't matter so much for normal clients, as they will retransmit
and cause the old request to be deleted from the request hash.
But detail requests have random ports (for other reasons), so
they won't be cleaned up by new packets.  Therefore, we need to clean
them up...