Fix weird issue where it would stop proxying packets.
If you had 4 servers, 1 of which was up and 3 down, it would
mark the 3 servers "dead", and then *erroneosly* mark the "live"
server as dead, too.

The symptom was that the home server would respond, but the proxy
would never see the response.  With some test configs from Bjorn Mork,
this was tracked down to b5b1ffbc7.  Since the "stable" branch didn't
have this problem, the solution was to revert packet_list_alloc()
to the pre-b5b... code, and then apply selective changes from the
"stable" branch