Delay decoding of packet until we are in the child thread
This also delays removing the packet from the proxy hash.  When the server
is busy, this means that the ID's are used for extended periods of time
as compared to 2.1.x.  However, if the server is busy and can't process
existing requests, then it has no business proxying even more packets
upstream.