More descriptive error messages.
If the thread pool queue is full, or if the packets are stuck in it for
a long time, it's usually because someone has a DB with 10 million rows
and no index.  Or, they're running long-lived queries against a MyISAM
database.  Either way, there's little that the RADIUS server can do
to fix it.

When that happens, produce a descriptive error message suggesting
that they focus the blame in the right place.