Various fixes
If timeout is zero, set it to 1^6.  That should be good enough.

Don't filter output through sed, this makes it impossible to kill the
"tail" process.

Catch the tail PID, and kill it on exit.

Trap on signals, not on normal exit.

Duplicate code in trap for normal exit.  This ensures that the code is
only executed once.