Added '-F' to filter RADIUS from input to output
This can be used to "clean up" horrible pcap files that people create.
i.e. when asked for a RADIUS pcap file, they just capture *everything*
on the interface for a period of time.  This makes it harder to track
down the real RADIUS issues.

Adding -F makes it easy to filter the packets.