Added "radattr" test tool.
This uses the parsing format (mostly) from the "extended attrs" draft.
It allows for automated testing of encode/decode of RADIUS attributes.

The current version doesn't use dictionaries.  Instead, it uses a brain-dead
parser to allow the tests to encode almost anything in almost any format.
This makes the encoder independent of any dictionary changes.

Future versions will read the dictionaries, and allow for encode/decode
of attributes.