Fixed "encode TLV" API.
We need to track the START vp, and the NEXT vp.  Previously,
we were tracking the *current* VP and the next one, which didn't
work.  The issue was that if the TLV encoding recursed twice,
it would not back up twice, and instead encode the next level 1
TLV as part of the current one.

Also updated the location of "debug_pair", to put it into
vp2data_any, which is really where it should go.