Fix autogen.sh
The script is invoked with -e which causes the script to exit
immediately if a subshell command enclosed in parentheses
has a non-zero exit status. The command

  grep "^AC_CONFIG_HEADER" configure.in > /dev/null

returns non-zero for many of the subdirs which causes the autogen.sh
script to immediately exit. In fact it exits on the very first
subdirectory (src/modules/rlm_sql) prematurely aborting the entire
operation.