Track and print time-based statistics
Now that we track packet->timestamp for requests and replies,
it can be used to keep track of client/global/home-server auth/acct
statistics for elapsed time.  We track the elapsed time
(reply - request) to see how long it takes for the server to respond
to requests.

We need to do the same thing for CoA packets, too.  Right now we
don't...