Handle relayed packets better...
If the request  a client packet, we can relay it using
the existing code.

If the request is a server packet, then it MUST be from
the real server, and we MUST be acting as a relay.  In that
case, set the giaddr to 0.0.0.0, and forward the packet to the
yiaddr.

And do something with broadcast replies...