Add undocumented "lazy init" configuration.
This allows the connection pool to return on init,
even if it's unable to make any new connections.  The result
is that the server can start even when the back-end DB is down.

That's nearly always a bad idea, but it's easy enough to do
with the new connection pool code.