Rearranged code to build with various compile options
WITHOUT_PROXY
WITHOUT_STATS

etc.

The server should build with any combination of the above options.
While they're not commonly used, they are useful.  When most
functionality is disabled, the server produces a ~600K statically
linked binary with all of the most common modules.

This is suitable for embedded systems...