Include copyright lines
This is for people who can't be bothered to give attribution
where it's due.  The dictionary files we have here give authors,
origins, web pages, etc. where possible.  It's only reasonable
for others to do the same.

Sadly, that's asking too much of some people.  So we've added
explicit copyright statements so that people have no excuse.