Rearrange packet list code
yank no longer returns a pointer.  No one was using it, so that
work was unnecessary.

Re-arrange the code in fr_packet_cmp() so that fewer comparisons
are necessary to disambiiguate packets.

Remove workaround for bug #35 in packet_entry_cmp().  It is
no longer necessary.

The result is a somewhat faster on in-memory performance tests.
But the callgrind output stil lshows large blocks of time
spent handling the packet lists.  Those could be optimized
some more.