Support new build system
After 10+ years, recursive make should be left behind.
This new framework is taken from:

	https://github.com/dmoulding/boilermake

with some changes by myself.  The basic framework is ~500 lines
of GNU Makefile code.  Adding libtool-style wrappers and install
frameworks is another ~500 LoC.  The result is that we now
have a build system which all developers should use.
Each individual Makefile is small, simple, etc.  Dependencies
are automatically calculated and tracked, so the developer
doesn't have to worry.

The minor downside is that you can't do "cd X;make".  This is
a GOOD thing, because Make now handles all dependencies.  i.e.
building in a subdirectory should NEVER be necessary.

All output goes into the "build" directory, which is a bit of
a change.