Add rate limiting for network sockets
via "max_pps" in the "listen" section.  It takes a count of the
packets received in the last second.  If it's over max_pps, the
new packet is ignored.  Otherwise, it's allowed.

We probably should instead have adaptive rate limiting based on
how many packets/s *finish* processing.  But that's harder to do
for now.