Hack dependencies to get "touch foo.c;make install" to work
the installed files depend on the "relink" versions, which have
the correct library paths in them.  The normal versions have the
in-source paths, to allow in-source execution.

The relink versions now have a dependency on the normal versions,
which causes the normal versions to get built, too.

We should probably change the "relink" target to "final", and
get rid of the ADD_LIBTOOL_PATH stuff.