Made packet data len "size_t".
Only one place needs it to be negative: the read from the socket.
For the rest of the code, it is ALWAYS positive.

As part of this, changed a few other variables, too.
And uses "%zu" for printing "size_t" variables.  C99 is everywhere.
If you want to build the server on a non-C99 compiler, go away.