Added preliminary support for EAP-Key-Name
We use EAP-Session-Id internally, as not everyone wants to send
EAP-Key-Name in a packet.  The eaptls_gen_eap_key() function
generates the EAP-Session-Id.  The eaptls_success() function
calls the "gen key" function on success.  This means that
the key is available on all TLS-based EAP methods

If someone wants to send EAP-Key-Name, they should use unlang
to make EAP-Session-Id to EAP-Key-Name