Make magic.
If config.log exists,
Grab the "configure" args from it,
and make the "*mk" files depend on "*.mk.in" and "configure".

If the *.mk.in file changes, or "configure" changes,
then re-run "configure" in that subdirectory, to re-generate
the *.mk file.

Right now it works with all.mk files, but not other *.mk files.
to be done later...