Fix horrible hack
The core of rlm_eap expects that "inst" is the structure
containing the parsed config variables.  Using "inst->conf"
instead confuses things.  And is a horrible idea.

So... we just put the conf stuff as the first structure in 'inst'
It's horrid, but simpler than going through and fixing it
correctly.