Make the "update" sections more generic.
The new template / map code does a ton of memory allocations,
but it's more generic than the old radius_update_attrlist()
function.

The template / map code largely duplicated the functionality
of radius_update_attrlist(), but with different code.
We've deleted the old code, and now use the new code for "update"
sections.  The benefit is that we can now do:

	update {
		reply:Filter-Id := &request:User-Name
	}

And it will work.  This is more generic than before.

As a bonus, the old 2.x "update" sections still work, too