Reverse the order of const and type specifier e.g. const char * becomes char const *
Style guides recommend this because if you read from right to left, it's much clearer
what is actually being marked as const.

const char * const *foo

Here the char array, and the pointer to the char array are const, but this isn't
immediately clear.

As the first const marks token to it's right as const, you'd expect the second const
to operate on the token to its right and make the char ** const, but it doesn't,
it instead marks the char * as const.

If we use the new format:

char const * const *foo

and read from right to left it's much clearer what's const and what's not.