Updated conditional parser to use new tmpl / map functions
And add normalization of conditions

This means that the "foo==bar" type of conditions are now maps.
So that both the LHS and RHS can be attribute references.

The next step is to add an "evaluate map / tmpl" function
to evaluate.c.  The conditions can then call it in order to
do all of the conditional checking