Add rlm_idn
This module will support any StringPrep needs especially IDNA.

It will be of limited practical usefulness until DDDS, which
requires it, is implemented, but at least this gets one small
separable task out of the way.

It is not simply added to rlm_expr for packaging reasons.
Despite the appearenceof simplicity, there's a lot of icky
unicode muckery, so it would be a bit onerous to provide an
in-tree implementation.

As such, for distros that split up rlms, this would allow severing
dependencies on libidn11 from the main freeradius package.