Be consistent with common config_item names
If a file is being referenced or created the config item 'filename' is used.

If a file is being created, the initial permissions are set by the 'permissions' config item.

If a directory hierarchy needs to be created the permissions are set by 'dir_permissions'

If an external host is referenced in the context of a module the 'server' config item is used.

file, script_file, module, detail, detailfile, attrsfile, perm, dirperm, detailperm,
hostname are all deprecated.