Add the 'debug_attr' xlat for power users
At debug lvl 2 it will print all attributes and values matching the attribute name. If a list is passed it will print the contents of the list.

At debug lvl 3 it will print out the attribute vendor, type and length.

At debug lvl 4 it will print out alternative decodings of the attribute value. This is useful when the dictionaries for a NAS are not available/the NAS is very broken. It also helps with debugging improperly typed attributes.

debug_attr: always expands to a zero length string.

When debugging is not enabled the xlat function exits immediately.