Don't delete the request if it's in the queue
request_done() should WAIT if the request is blocked in the queue.

It should also be callable by the child thread.  i.e. if the
master says "DONE" when it's in the queue, the child thread MAY
pick it up and call request->process().  Which should just say DONE
and stop