Fix clang build under mavericks (OSX 10.9)
The default header files have moved, so they're no longer
in /usr/include/.  <sigh>

So we do special checks to find them.  And then add the special
sauce to jlibtool as DARWIN_CFLAGS, so it doesn't pick up the
rest of the CFLAGS we're using to build the server.

We also add the sauce to the normal CFLAGS and LDFLAGS, too,
so the normal build can proceed.

BUT, we add the special sauce to the END of the cflags.  So if
you want to use locally built libraries instead of the system
ones (I'M LOOKING AT YOU, OPENSSL), you can.