Extend the H3C dictionary based on public manuals
Try to extend the H3C dictionary based on what is scattered on
their pages. At least most of it is found quite consistently in
the product-specific "Security Configuration Guides".

This compilation/crosschecking is based H3C documentation for
listed devices:
- H3C S12500 and S6800 Switches
- H3C WA WLAN Access Points
- H3C MSR Routers

Remarks:
Does not stick with H3C's use of of underline in attribute naming.