Revert "Auto-load dictionary.vqp and dictionary.dhcp"
This reverts commit 013b841fadc72844642c559201d7e475374826e5.

We'll do this later.

We need to update mainconfig.c to set dictionary_dir to the
correct place.  It should load the default dictionaries
first.  And then try to load the dictionary from raddb, with
a soft-fail if it doesn't exist.

dict.c needs to be updated so that it stats() the file prior
to loading it.  This ensures that the above change is fail-safe.
i.e. if the raddb/dictionary file still has a $INCLUDE of the
default ones, the entries aren't re-loaded in their entirety