Make TCP sockets work with udpfromto.
Add a "proto" field to RADIUS_PACKET.
Have it checked in rad_send(), so that we just use write().
Using sendto() is OK for TCP sockets.
Using udpfromto is not OK for TCP sockets.

Ensure that the TCP code in listen.c sets the proto field.
Ensure that rad_alloc_reply() sets the proto field
Ensure that radclient sets the proto field as necessary

Conflicts:
	src/main/radclient.c