Add missed header check for features.h
Apparently nothing else checks for GNU versions of functions.

The libc ubuntu 12.04 ships with has the GNU version of strerror_r, whereas OSX has the Posix one. The compiler should of emitted a warning, as the return type of the functions differed, but didn't, for reasons best known to it.