Remove templating from cf_pair_find().
This breaks the server, so the next commit will add templates
back in again, in a better fashion.

We also remove the ability to do cf_pair_find(cs, NULL), and have
it return the first CONF_PAIR.  After searching the code, no one
uses this functionality.  So it's best to remove it.