Fix templating code.
Instead of hacking a reference into cf_pair_find() but not
cf_section_*_find(), we just copy the template recursively.

When we reach the end of a section with a template, we merge
a copy of the template into the current section, then we nuke
the relationship between the section and the template.

This uses more memory, but it means that templates now work
with subsections