Fix command parser and debug_condition.
The command parser no longer crushes "foo" --> foo
Instead, it leaves everything as-is.  This shouldn't affect
anything but the debug condition.  All other commands should
not be taking quoted strings.

Fix the debug_condition code to use *all* of it's arguments.
(User-Name=="bob") was OK.  (User-Name == "bob") was parsed
as 3 arguments, and the second 2 were ignored