Add "secret" flag to configuration items.
For passwords, shared secrets, etc.
For now, strings only.

When using "radiusd -X", secret items are NOT printed,
and are replaced with <<< secret >>>

They are still printed out when using "radiusd -Xx".
So administrators who care to know can still find them.

The User-Password is *not* mangled in this manner.  We have
too many people with shared secret problems in the list.  It is
useful to be able to point to the debug output and say
"LOOK, it's WRONG".

But passwords for LDAP, SQL, etc. are another class of object.
So it's OK to hide them in "radiusd -X"