Don't fill up the log with per-packet messages.
When we can't open a new outgoing proxy socket, that's an error,
which is now rate-limited to once per second.

When we can't add a request to the proxy socket as a result of
the previous error, just print an EDEBUG message.  Don't do
RPROXY, which causes it to be written to the log file.