Load protocol-specific dictionaries if built with them.
The code in parser.c does static evaluation of things like

	if (DHCP-foo == ...)

When the DHCP dictionaries aren't loaded, that turns into static
comparison of two strings, which is always false.  Changing it to

	if (&DHCP-foo == ...)

makes it bail with an unknown attribute error.  This is because
we don't have pass2 fixups for that case.