Parsing 'foo%%s' requires an intermediate literal of "%"
The %% takes 2 characters. We need one to terminate the
previous string.  One for the %.  And one to terminate
the "%" string.  That's three, which doesn't fit into the
space of %%.  So... talloc_strdup to the rescue

We could add "optimization" code to detect of the previous
character was zero, OR the previous node was a literal,
and avoid the creation of a new node / talloc.  But... that's
more fragile.  Just doing the simple thing will work